Combined Login and Sign Up pages
* Removed Sign Up page
