--- CarCamApp/Views/SignUpPage.xaml
@@ -1,25 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:d="http://xamarin.com/schemas/2014/forms/design"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             mc:Ignorable="d"
-             x:Class="CarCamApp.Views.SignUpPage">
-
-
-    <StackLayout>
-        <StackLayout>
-            <Label Text="You are signing up for an account on the Car Cam App!" HorizontalOptions="CenterAndExpand" FontSize="Title"/>
-        </StackLayout>
-
-
-        <StackLayout>
-            <Label x:Name="Lbl_Username" Text="Username:"/>
-            <Entry x:Name="Entry_Username" Placeholder="Username"/>
-            <Label x:Name="Lbl_Password" Text="Password:"/>
-            <Entry x:Name="Entry_Password" Placeholder="Password" IsPassword="True"/>
-            <Button x:Name="Btn_SignUp" Text="Sign Up" HorizontalOptions="End" Clicked="SignUpProcedure"/>
-    </StackLayout>
-        
-    </StackLayout>
-</ContentPage>
